Support resetting metrics and flag for appending hostname.
In some cases it is preferable to have metrics be reset after
submitted to graphite. This allows metrics in graphite to correctly
reflect the values for the time period only.

folsom_graphite normally adds the hostname to the metric name. This
is almost always desirable, but not always. This change makes this
behavior configurable.